remove module index indirection
